# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorMessage {
  error: string;
}

export interface Item {
  id: string;
  name: string;
  category: string;
  quantity: number;
  unit: string;
  minStock: number;
  /** @nullable */
  cost?: number | null;
  /**
     * Current inventory value (cost multiplied by quantity)
     * @nullable
     */
  totalValue?: number | null;
  addedBy: string;
  addedAt: string;
  updatedAt: string;
}

export interface ItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  quantity: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  minStock: number;
  /** @nullable */
  cost?: number | null;
  /** @minLength 1 */
  performedBy: string;
}

export interface ItemUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  /** @minimum 0 */
  quantity?: number;
  /** @minLength 1 */
  unit?: string;
  /** @minimum 0 */
  minStock?: number;
  /** @nullable */
  cost?: number | null;
  /** @minLength 1 */
  performedBy: string;
}

export interface ItemDeletion {
  /** @minLength 1 */
  performedBy: string;
}

export interface StockAction {
  /** @exclusiveMinimum 0 */
  quantity: number;
  /** @minLength 1 */
  performedBy: string;
  note?: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  restock: 'restock',
  withdrawal: 'withdrawal',
  adjustment: 'adjustment',
  deletion: 'deletion',
  addition: 'addition',
} as const;

export interface Transaction {
  id: string;
  itemId: string;
  itemName: string;
  type: TransactionType;
  quantityBefore: number;
  quantityAfter: number;
  delta: number;
  performedBy: string;
  timestamp: string;
  /** @nullable */
  note?: string | null;
}