# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Item {
  id: string;
  name: string;
  category: string;
  quantity: number;
  unit: string;
  minStock: number;
  /** @nullable */
  cost?: number | null;
  /**
     * Current inventory value (cost multiplied by quantity)
     * @nullable
     */
  totalValue?: number | null;
  addedBy: string;
  addedAt: string;
  updatedAt: string;
}