# lib/api-zod/src/generated/types/itemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  quantity: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  minStock: number;
  /** @nullable */
  cost?: number | null;
  /** @minLength 1 */
  performedBy: string;
}