# lib/api-zod/src/generated/types/stockAction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StockAction {
  /** @exclusiveMinimum 0 */
  quantity: number;
  /** @minLength 1 */
  performedBy: string;
  note?: string;
}