# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from './transactionType';

export interface Transaction {
  id: string;
  itemId: string;
  itemName: string;
  type: TransactionType;
  quantityBefore: number;
  quantityAfter: number;
  delta: number;
  performedBy: string;
  timestamp: string;
  /** @nullable */
  note?: string | null;
}